requirements.txt
# Vector Database
chromadb==1.0.20

# LLM Integration
openai==1.101.0

# Environment Variables
python-dotenv==1.0.0

# API Framework
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic>=2.0.0
